Move begin step into step package
