+++ hub/initialize.go
@@ -18,7 +18,7 @@ import (
 )
 
 func (s *Server) Initialize(in *idl.InitializeRequest, stream idl.CliToHub_InitializeServer) (err error) {
-	st, err := BeginStep(s.StateDir, "initialize", stream)
+	st, err := step.Begin(s.StateDir, "initialize", stream)
 	if err != nil {
 		return err
 	}
@@ -46,7 +46,7 @@ func (s *Server) Initialize(in *idl.InitializeRequest, stream idl.CliToHub_Initi
 }
 
 func (s *Server) InitializeCreateCluster(in *idl.InitializeCreateClusterRequest, stream idl.CliToHub_InitializeCreateClusterServer) (err error) {
-	st, err := BeginStep(s.StateDir, "initialize", stream)
+	st, err := step.Begin(s.StateDir, "initialize", stream)
 	if err != nil {
 		return err
 	}
